tr_shader.cpp: fix errors on GCC8
GCC8 refuses to compile this file because of implicit truncation between
integer variables and constants. Extend the size of the in-memory table to fix the issue.

+++ code/renderer/tr_shader.cpp
@@ -1854,7 +1854,7 @@ void ParseSort( const char **text )
 
 typedef struct {
 	char	*name;
-	int		clearSolid, surfaceFlags, contents;
+	long int		clearSolid, surfaceFlags, contents;
 } infoParm_t;
